DP 329h Longest Increasing Path in a Matrix
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.81s
     Running unittests src/lib.rs (target/debug/deps/dp-8ed43879319390e2)

running 1 test
* [[9, 9, 4], [6, 6, 8], [2, 1, 1]]
-> (0, 0)
-> . (0, 1)
-> . (1, 0)
-> (0, 1)
-> . (0, 2)
-> . (1, 1)
-> . (0, 0)
-> (0, 2)
-> . (1, 2)
-> (1, 2)
-> . (0, 2)
-> . (2, 2)
-> . (1, 1)
-> . (0, 1)
-> (1, 0)
-> . (0, 0)
-> . (1, 1)
-> . (2, 0)
-> (1, 1)
-> . (0, 1)
-> . (1, 2)
-> . (2, 1)
-> . (1, 0)
-> (2, 0)
-> . (1, 0)
-> . (2, 1)
-> (2, 1)
-> . (1, 1)
-> . (2, 2)
-> . (2, 0)
-> (2, 2)
-> . (1, 2)
-> . (2, 1)
-> [[1, 1, 2], [2, 2, 1], [3, 4, 2]]
:: 4
* [[3, 4, 5], [3, 2, 6], [2, 2, 1]]
-> (0, 0)
-> . (0, 1)
-> (0, 1)
-> . (0, 2)
-> (0, 2)
-> . (1, 2)
-> (1, 2)
-> . (0, 2)
-> . (2, 2)
-> . (1, 1)
-> . (0, 1)
-> . (1, 1)
-> . (0, 0)
-> . (1, 0)
-> (1, 0)
-> . (0, 0)
-> . (1, 1)
-> . (2, 0)
-> (1, 1)
-> . (0, 1)
-> . (1, 2)
-> . (2, 1)
-> . (1, 0)
-> (2, 0)
-> . (1, 0)
-> . (2, 1)
-> (2, 1)
-> . (1, 1)
-> . (2, 2)
-> . (2, 0)
-> (2, 2)
-> . (1, 2)
-> . (2, 1)
-> [[4, 3, 2], [1, 4, 1], [2, 1, 2]]
:: 4
* [[1]]
-> (0, 0)
-> [[1]]
:: 1
test tests::test_329 ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 33 filtered out; finished in 0.00s

+++ dp/src/lib.rs
@@ -212,7 +212,7 @@ impl Sol329 {
         println!("-> {dp:?}");
 
         if let Some(&max) = dp.iter().flatten().max() {
-            return max as i32;
+            return max;
         }
         1
     }
